Refatoração estrutural dos veiculos
