+++ client/src/components/InfoVeiculo/index.js
@@ -1,5 +1,5 @@
 import React, { Component } from 'react'
-import CarIcon  from '../ListVeiculo/veiculo_preto.png'
+import CarIcon  from '../Veiculo/ListVeiculo/veiculo_preto.png'
 
 export default class InfoVeiculo extends Component {
   render() {
