+++ client/src/components/LiftPend/index.js
@@ -8,7 +8,7 @@ import CadeiranteIcon from '../LiftMgt/cadeirante_roxo.png'
 import FumanteIcon from '../LiftMgt/fumante_roxo.png'
 import SocketIOChat from '../SocketIOChat'
 import MusicIcon from '../LiftMgt/musica_roxo.png'
-import CarIcon from '../ListVeiculo/veiculo_preto.png'
+import CarIcon from '../Veiculo/ListVeiculo/veiculo_preto.png'
 //import LugarIcon from '../Veiculo/lugares_preto.png'
 import { loadCarbyID } from '../../actions/carActions'
 import { insertAvalicao, delCaronaPendMotorista, delCaronaPendCaronista } from '../../actions/liftActions'
