+++ client/src/components/Menu/menuContent.js
@@ -32,7 +32,7 @@ const itens =
     },
     {
         text: "Meus Veiculos",
-        route: "/"
+        route: "/veiculos"
     },
     {
         text: "Notificações",
