+++ client/src/components/Veiculo/ListVeiculo/lugares_preto.png
