+++ client/src/components/Veiculo/ListVeiculo/lugares_roxo.png
