+++ client/src/components/Veiculo/ListVeiculo/veiculo_branco.png
