+++ client/src/components/Veiculo/ListVeiculo/veiculo_preto.png
