+++ client/src/components/Veiculo/detalhes/style.css
@@ -1,0 +1,3 @@
+.veiculos-detalhe .icon {
+    width: 200px;
+}
