+++ client/src/pages/oferecer/index.js
@@ -7,7 +7,7 @@ import TimeIcon from 'material-ui/svg-icons/action/schedule'
 import CadeiranteIcon from '../../components/LiftMgt/cadeirante_branco.png'
 import SmokingIcon from '../../components/LiftMgt/fumante_branco.png'
 import MusicIcon from '../../components/LiftMgt/musica_branco.png'
-import CarIcon from '../../components/ListVeiculo/veiculo_branco.png'
+import CarIcon from '../../components/Veiculo/ListVeiculo/veiculo_branco.png'
 import { buscarRotas } from '../../actions/mapActions'
 import { loadCar } from '../../actions/carActions'
 import { offerLift } from '../../actions/liftActions'
